requirements.txt
python-dotenv
openai
langchain
langchain-openai
langchain-anthropic